pythonProject/book_service/requirements.txt

amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.4
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
django-import-export==3.3.7
django-oauth-toolkit==2.3.0
django-redis==5.4.0
django-rest-framework-social-oauth2==1.2.0
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.11.0
flower==2.0.1
gunicorn==22.0.0
humanize==4.9.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kombu==5.3.7
MarkupPy==1.14
model-bakery==1.17.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
tablib==3.5.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0


django-baton